Move ProgressScreenFeature to progress_screen package (#562)
